requirements.txt
openai
langchain
faiss-cpu
numpy
pandas
flask
tiktoken
langchain_community
sentence_transformers
google-generativeai
